Backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pandas==2.2.0
openai==1.12.0
python-dotenv==1.0.0
alembic==1.13.1
backend==0.1.0
werkzeug>=3.0.0